backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pyswisseph>=2.10.0
tzdata>=2024.1
